requirements.txt
# gRPC相关依赖
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# JWT认证
PyJWT==2.8.0

# 现有依赖（Flask和数据库）
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
Werkzeug==3.0.1

# 文件编码检测
chardet==5.2.0

# 数据处理
cryptography==41.0.7
